+++ collections/composers/persistable.d.ts
@@ -1,4 +1,0 @@
-export function persistable({ beforeInsert, beforeUpdate }?: {
-    beforeInsert?: Function;
-    beforeUpdate?: Function;
-}): Function;