+++ collections/composers/softRemoval.d.ts
@@ -1,1 +1,0 @@
-export function softRemoval(collection: Mongo.Collection): any;