+++ collections/schema.d.ts
@@ -1,1 +1,0 @@
-export function extendSchema(s1: any, s2: any): any;