+++ collections/CHANGELOG.md
@@ -1,5 +1,14 @@
 ## CHANGELOG
 
+### 3.0.4 (2024-10-13)
+
+- Adds types using only zodern:types.
+
+### 3.0.3 (2024-10-13)
+
+- Adds `isServerOnly` on `createCollection` function as an option.
+- Adds tests.
+
 ### 3.0.2 (2024-10-10)
 
 - Adds `persistable` composer